New off-line signing flow with function options for submit and evaluate
The stateless off-line signing flow is now:
- `contract.NewProposal()` with function options
- Return `proposal.Bytes()` and `proposal.Hash()` to off-line signer
- Receive proposal and signature bytes from off-line signer
- `contract.NewSignedProposal()`
- `proposal.Endorse()` to get an endorsed `Transaction`
- Return `transaction.Bytes()` and `transaction.Hash()` to off-line signer
- Receive transaction and signature bytes from off-line signer
- `contract.NewSignedTransaction()`
- `transaction.Submit()`

Function options seem to make submit and evaluate so simple that the long and short forms look almost identical, so just the long form is retained for now.

Function options currently apply to all proposals and don't distinguish between evaluate and submit flows. This is OK for arguments and transient data as they apply to both flows. As we discover options that are only applicable to one of these flows, the function options will need to be typed to only apply to either an evaluate or submit.