Stop receiving chaincode events in Java API
Add a close() method to the event iterator that can be called when events are no longer required.

Calling close() on the event iterator cancels the server streaming gRPC connection from the client end. This requires the use of the async gRPC stub, which is significantly more complex to use than the blocking API. This complexity is hidden within a new internal GatewayClient class, which decouples the rest of the client code from the gRPC API calls.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>