Remove ChaincodeEventRequest.onEvent() method from Node client
Callback style eventing can be easily implemented by a client application on top of the async iterator API. This avoids the need to implement closure of the eventing session in the callback style API, where the optimum implementation is less obvious. This can be added later if there is demand.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>