Update Fabric protobuf definitions
Use a version of the protobuf definitions that include the `after_transaction_id` field in the ChaincodeEventsRequest message. This allows the client to resume chaincode eventing after a specfic (previously processed) transaction within a block, and removes the need for checkpointer logic to filter previously processed events at the client end.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>