Use OptionalLong for Java checkpoint block number
Block number will not be set in the initial checkpointer state, and
previously logic to detect this required that the block number be set to
zero and the transaction ID (which already uses an Optional) unset. This
can be simplified by using an Optional to indicate whether the block
number has a value.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>