Require Go gRPC client connection interface for Gateway connect
The code previously required the concrete client connection type, which is likely a hang-over from early development of the API where Gateway instances might call Close() on the concrete type. The current design deliberately leaves management of the gRPC connection to the caller so only requires the client connection interface used by gRPC service stubs. This is a non-breaking change since the interface is a subset of the methods provided by the concrete type.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>