Increment minor version for new functionality
The Go API was extended so incrementing the minor version to v1.2.0.

As documented, each minor version supports only the currently supported language versions of Go, Node and Java. Drop support for Go 1.17 since this is no longer a current Go version, and use Go 1.18 as the minimum version.

Update dependencies.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>