Use staticcheck binary for linting
The staticcheck linter in golangci-lint is not the real staticcheck binary. Disable the staticcheck linter in golangci-lint and use the real staticcheck binary, just to ensure the latest rules are being applied.

Previous refactor of build jobs ensures that linting (including staticcheck) is only run with a Go version supported by staticcheck.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>