Include TLS configuration in code examples
For simplicity the code examples in the API documentation created insecure gRPC connections. This led to some confusion for users when the code examples did not work with a Fabric network configured for TLS, which is the typical Fabric configuration.

This change updates the example code snippets in the API documentation to include explicit TLS configuration of the gRPC connection.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>