Scenario test oldest and newest language versions
Save build cycles by only running scenario tests using the oldest and newest versions of each of the supported languages. For example, Java scenario tests are run with only Java 8 and Java 21, not Java 11 or Java 17. Unit tests are much cheaper so are still run with all supported language versions.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>