Include missing font resources in JavaDoc
Site content generated by the JavaDoc tool references
`resources/fonts/dejavu.css` in its CSS stylesheet. This file (and the
fonts that it references are missing from the generated output. This is
a known JavaDoc bug.

This change includes the missing content as resources in the project
source, and configures the Maven JavaDoc plugin to include them in the
JavaDoc output. This results in the correct fonts being used in the
generated JavaDoc, matching the style of the official Java API
documentation.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>