Use latest stable Go and Node releases in CI
For GitHub Actions jobs that are not sensitive to the version of Go or
Node used, use "stable" or "lts/*" as the version for Go and Node
respectively.

Previously the version number was specified explicitly, which requires
more ongoing maintenance and conveys less clarity of the intent that the
latest stable version should be used.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>