Fix GatewayException.printStackTrace(PrintStream)
The printStackTrace() implementation taking a PrintStream (rather than a
PrintWriter) in the Java GatewayException class failed to print output.
This was due to the PrinterWriter used to wrap the supplied PrintStream
not being flushed.

This change calls the PrintStream overload with unit tests, instead of
the PrintWriter overload (to which it defers) to ensure both variants
behave correctly. The printStackTrace() implementation is updated to
correctly flush the PrintWriter used internally to wrap the supplied
PrintStream.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>