Include error details in unchecked Java exception (#768)
The checked GatewayException already has its printStackTrace augmented
with any associated gRPC ErrorDetail elements. The same functionality is
added to the equivalent unchecked GatewayRuntimeException. The two
exceptions are intended to behave identically, with the unchecked
variant to be used only where a checked exception cannot be thrown due
to the standard APIs exposed. For example, the Iterator methods used
when retrieving events.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>