Fix mockery install to work on Apple Silicon Macs
The URL used to download the Mockery binary is composed using the
current hardware platform string. For Apple Silicon Macs, this is
`aarch64`, whereas the Mockery binary uses `arm64`.

This change uses `arm64` in place of `aarch64` as the hardware platform
to build the Mockery download URL.

The version of Mockery is also updated.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>