Avoid use of npm-run-all in npm scripts (#787)
npm-run-all is used as a convenience to run multiple npm scripts
sequentially. This can also be achieved by simply chaining shell
commands. For example: `npm run a && npm run b`. One of the packages in
the npm-run-all dependency tree has a reported security vulnerability,
and npm-run-all seems to be unmaintained having not published a release
for 6 years.

This change removes the use of npm-run-all as a convenience and instead
uses shell commands to sequentially invoke npm scripts.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>