Change surefire naming strategy
The `surefire` naming strategy for the Java surefire test runner plugin
does not work as expected for surefire plugin 3.5.4 and above. This
causes warning messages to appear in the test logs, and can compromise
the usability of the output.

This change use the currently recommended `long` naming strategy. There
are some additional minor changes to dependency declarations and
docker-compose files to streamline version management and avoid warning
messages during test execution.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>